requirements.txt
# TrackImmo Backend Dependencies

# Web Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2

# Database
sqlalchemy==2.0.21
# psycopg2-binary==2.9.7  # Uncomment for production with PostgreSQL
# For development, use sqlite
alembic==1.12.0

# Data Processing
pandas==2.1.0
numpy==1.25.2

# HTTP Client
requests==2.31.0
httpx==0.24.1

# Scraping
playwright==1.39.0
beautifulsoup4==4.12.2

# Data Validation
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
loguru==0.7.2
tqdm==4.66.1
prometheus-client==0.17.1

# Geographic Data
geopy==2.4.0
shapely==2.0.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.24.1 